backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
oauthlib==3.2.2
requests==2.28.2
python-dotenv==0.21.0 
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0 
psycopg2==2.9.5